erasure-code: Move the EC factory code into separate header files
This helps the EC factories be invokable directly for testing
and other uses.
