+++ src/erasure-code/jerasure/ErasureCodePluginJerasure.cc
@@ -17,8 +17,8 @@
 
 #include "ceph_ver.h"
 #include "common/debug.h"
-#include "erasure-code/ErasureCodePlugin.h"
 #include "ErasureCodeJerasure.h"
+#include "ErasureCodePluginJerasure.h"
 #include "jerasure_init.h"
 
 #define dout_subsys ceph_subsys_osd
@@ -30,9 +30,7 @@ static ostream& _prefix(std::ostream* _dout)
   return *_dout << "ErasureCodePluginJerasure: ";
 }
 
-class ErasureCodePluginJerasure : public ErasureCodePlugin {
-public:
-  virtual int factory(const std::string& directory,
+int ErasureCodePluginJerasure::factory(const std::string& directory,
 		      ErasureCodeProfile &profile,
 		      ErasureCodeInterfaceRef *erasure_code,
 		      ostream *ss) {
@@ -70,8 +68,7 @@ public:
     }
     *erasure_code = ErasureCodeInterfaceRef(interface);
     return 0;
-  }
-};
+}
 
 const char *__erasure_code_version() { return CEPH_GIT_NICE_VER; }
